Split refactored Arena into BubbleStorage and Collision, now properly working
